Gitlab/Gitolite: Support anonymous access through the git daemon
This adds the option for enabling and disabling anonymous access to a
repo via adding the pseudo-user "daemon" to gitolite's access list.

If anonymous Git access is enabled, GitLab will display the access link
on the repo.

Obviously, your GitD must be configured with the correct base directory
and execute as the appropriate user (probably git).

Conflicts:
	app/models/project.rb
	app/models/repository.rb
	app/views/admin/projects/show.html.haml
	db/schema.rb